Add basic material creation from a property list
So far this differs from instantiating a material directly only so far as being
able to cast values from strings to a more appropriate type for a certain subset
of properties. This currently excludes complex properties (like clipping planes)
and anything involving or relating to texture maps (which would require extra
handling of `url(...)` properties and async loading).